Change extension to lowercase
The file was previously named "chocolate-cake.JPG". I locally changed it
to lowercase because I think that's better practice. I also changed the
file path in chocolate-cake.html to "../img/chocolate-cake.jpg".

When I viewed chocolate-cake.html locally, the image loaded correctly.
When I viewed the page hosted on Github, the image was broken. I think
this is because the image name on Github was still "chocolate-cake.JPG".